Probably should have started recording earlier
I tried to just mess around with vectors, and I found a way to have a
single object be attracted to a stationary object. I found the idea
quite amusing at the time, so I decided to develop on it. My first goal
was to see how adding more stationary objects would change the moving
one, which was amusing for some time, but got tired quickly.
Eventually, I decided to remove the stationary objects all together and
added another moving object. That was good and all, but then I decided
to incorporate controls into it, which made it a lot more fun. Adding a
little bit of flair to it with some sprites and colouring, and I got a
pretty interesting concept. I plan to add some kind of simple shooting
mechanic in the future, which since I can’t use a mouse, should
probably be a simple homing type of shot.

+added attracting objects
+added player interactivity
+added art style